// src/app/@core/interfaces/auction/affair-sera.model.ts
export interface AffairSeraInterface {
    code: number | null;
    description: string | null;
    relationPropertyKey: string | null;
    referralNoteType: string | null;
    versionUser: string | null;
    creationUser: string | null;
    creationDate: Date | null;
    editionUser: string | null;
    modificationDate: Date | null;
    idRegister: number | null;
  }